initial code for H3 Index

+++ rebuild.sh
@@ -1,6 +1,8 @@
 #! bin/bash
-rm /home/adam/milvus/cmake_build/bin/geos-config
-make -j8
+if [[ "$1" != "--ignore" ]]; then
+    rm /home/adam/milvus/cmake_build/bin/geos-config
+    make -j8
+fi
 pushd deployments/docker/dev/
 docker compose down
 sudo rm -rf ./volumes/*
